+++ src/yuzu/CMakeLists.txt
@@ -43,6 +43,8 @@ add_executable(yuzu
     game_list.cpp
     game_list.h
     game_list_p.h
+    game_list_worker.cpp
+    game_list_worker.h
     hotkeys.cpp
     hotkeys.h
     main.cpp
